+++ src/main/kotlin/uk/gov/dluhc/printapi/config/LoggingConfiguration.kt
@@ -2,9 +2,11 @@ package uk.gov.dluhc.printapi.config
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import uk.gov.dluhc.logging.config.CorrelationIdMdcInterceptor
 import uk.gov.dluhc.logging.config.CorrelationIdMdcMessageListenerAspect
 import uk.gov.dluhc.logging.config.CorrelationIdMdcScheduledAspect
+import uk.gov.dluhc.logging.rest.CorrelationIdMdcInterceptor
+import uk.gov.dluhc.logging.rest.CorrelationIdRestTemplateClientHttpRequestInterceptor
+import uk.gov.dluhc.logging.rest.CorrelationIdWebClientMdcExchangeFilter
 
 @Configuration
 class LoggingConfiguration {
@@ -17,4 +19,11 @@ class LoggingConfiguration {
 
     @Bean
     fun correlationIdMdcScheduledAspect() = CorrelationIdMdcScheduledAspect()
+
+    @Bean
+    fun correlationIdRestTemplateClientHttpRequestInterceptor() =
+        CorrelationIdRestTemplateClientHttpRequestInterceptor()
+
+    @Bean
+    fun correlationIdWebClientMdcExchangeFilter() = CorrelationIdWebClientMdcExchangeFilter()
 }
