+++ src/main/kotlin/uk/gov/dluhc/printapi/config/WebClientConfiguration.kt
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.reactive.function.client.WebClient
+import uk.gov.dluhc.logging.rest.CorrelationIdWebClientMdcExchangeFilter
 
 @Configuration
 class WebClientConfiguration {
