+++ src/main/kotlin/uk/gov/dluhc/printapi/config/WebMvcConfig.kt
@@ -3,7 +3,7 @@ package uk.gov.dluhc.printapi.config
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-import uk.gov.dluhc.logging.config.CorrelationIdMdcInterceptor
+import uk.gov.dluhc.logging.rest.CorrelationIdMdcInterceptor
 
 @Configuration
 class WebMvcConfig(private val correlationIdMdcInterceptor: CorrelationIdMdcInterceptor) : WebMvcConfigurer {
